feat(auth): simplify to single admin user model
Remove multi-user management (list, delete, toggle admin) in favour of
a single always-admin user.  Only allowed self-service operation is
changing own password via PUT /user/password with current-password
verification.

Backend:
- Remove ListUsers, DeleteUser, SetAdminStatus, UpdateUser from UserRepository
- Replace handleListUsers/handleUpdateUserAdmin/handleAdminChangeUserPassword
  with handleChangeOwnPassword
- Remove /admin group and legacy /users/* routes from server
- Fix arrs/config_handlers to use GetAllUsers instead of ListUsers
- Skip OAuth user profile update (UpdateUser) on login

Frontend:
- Delete UserManagement component and /admin route
- Remove Users nav item from Sidebar
- Add Change Password modal to UserMenu (current + new + confirm)
- Replace getUsers/updateUserAdmin/changeUserPassword API methods
  with changeOwnPassword (PUT /user/password)
- Replace UserAdminUpdateRequest/ChangeUserPasswordRequest types
  with ChangeOwnPasswordRequest
