+++ frontend/src/components/auth/index.ts
@@ -2,5 +2,4 @@ export { DirectLoginForm } from "./DirectLoginForm";
 export { LoginPage } from "./LoginPage";
 export { ProtectedRoute } from "./ProtectedRoute";
 export { RegisterForm } from "./RegisterForm";
-export { UserManagement } from "./UserManagement";
 export { UserMenu } from "./UserMenu";
