+++ frontend/src/components/layout/Sidebar.tsx
@@ -9,7 +9,6 @@ import {
 	Home,
 	List,
 	Settings,
-	Users,
 } from "lucide-react";
 import { NavLink } from "react-router-dom";
 import { useHealthStats, useQueueStats } from "../../hooks/useApi";
@@ -35,12 +34,6 @@ const navigation = [
 		name: "Files",
 		href: "/files",
 		icon: Folder,
-	},
-	{
-		name: "Users",
-		href: "/admin",
-		icon: Users,
-		adminOnly: true,
 	},
 	{
 		name: "Configuration",
